Add shuffle functionality to DeckOfCards
Added a shuffle method to the deck of cards. This method makes use
of shuffling algorithm that can be replaced.
Added ShufflingAlgorithm interface with a single shuffle method.
Added the ShufflingAlgorithm instance variable to the DeckOfCards
class which can be set using the constructor.
Added the default shuffling algorithm implementation which
uses the Collections.shuffle() method.
Added a method DeckOfCards.getCards() to get the cards of a deck.